Autorização

+++ app/Http/Controllers/ProjectController.php
@@ -6,6 +6,7 @@ use CodeProject\Http\Requests;
 use CodeProject\Repositories\ProjectRepository;
 use CodeProject\Services\ProjectService;
 use Illuminate\Http\Request;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 
 class ProjectController extends Controller
 {
@@ -27,7 +28,6 @@ class ProjectController extends Controller
      */
     public function __construct(ProjectRepository $repository, ProjectService $service)
     {
-
         $this->repository = $repository;
         $this->service = $service;
     }
@@ -40,7 +40,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id'=>Authorizer::getResourceOwnerId()]);
     }
 
     /**
@@ -62,7 +62,12 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
-        return $this->repository->with(['owner', 'notes', 'client'])->find($id);
+        if($this->checkProjectPermissions($id)==false){
+            return [
+                'msg'=>'Forbidden'
+            ];
+        }
+        return $this->repository->with(['owner', 'notes', 'client', 'members'])->find($id);
     }
 
 
@@ -86,6 +91,35 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if($this->checkProjectOwner($id)==false){
+            return [
+                'msg'=>'Forbidden'
+            ];
+        }
+
         $this->repository->delete($id);
+    }
+
+    private function checkProjectOwner($projectId){
+        $userId = Authorizer::getResourceOwnerId();
+
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+
+    private function checkProjectMember($projectId){
+
+        $userId = Authorizer::getResourceOwnerId();
+
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    private function checkProjectPermissions($projectId){
+
+        if($this->checkProjectOwner($projectId) OR $this->checkProjectMember($projectId)){
+            return true;
+        }
+
+        return false;
     }
 }
